+++ react/components/GradientCollapse/index.js
@@ -38,7 +38,7 @@ function GradientCollapse(props) {
     collapsed: collapsedProp,
   } = props
 
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useCssHandles(CSS_HANDLES)
   const [collapsed, setCollapsed] = useState(collapsedProp)
   const [prevCollapsedProp, setPrevCollapsedProp] = useState(collapsedProp)
   const [maxHeight, setMaxHeight] = useState('auto')